refactor: extract base issue logic from issue-detail-layout to logic/base.ts
