+++ frontend/src/components/Issue/IssueDetailLayout.vue
@@ -69,24 +69,7 @@
             <div
               class="py-6 lg:pl-4 lg:w-96 xl:w-112 lg:border-l lg:border-block-border overflow-hidden"
             >
-              <IssueSidebar
-                :issue="issue"
-                :database="database"
-                :instance="instance"
-                :create="create"
-                :selected-stage="selectedStage"
-                :task="selectedTask"
-                :input-field-list="issueTemplate.inputFieldList"
-                :allow-edit="allowEditSidebar"
-                :is-tenant-deploy-mode="isTenantMode"
-                @update-assignee-id="updateAssigneeId"
-                @update-earliest-allowed-time="updateEarliestAllowedTime"
-                @add-subscriber-id="addSubscriberId"
-                @remove-subscriber-id="removeSubscriberId"
-                @update-custom-field="updateCustomField"
-                @select-stage-id="selectStageOrTask"
-                @select-task-id="selectTaskId"
-              />
+              <IssueSidebar :database="database" :instance="instance" />
             </div>
             <div class="lg:hidden border-t border-block-border" />
             <div class="w-full py-4 pr-4">
@@ -121,20 +104,13 @@
 /* eslint-disable vue/no-mutating-props */
 
 import { computed, onMounted, PropType, ref, watch, watchEffect } from "vue";
-import { useRoute, useRouter } from "vue-router";
-import { cloneDeep, isEmpty, isEqual } from "lodash-es";
+import { useRoute } from "vue-router";
 import {
-  idFromSlug,
   pipelineType,
   PipelineType,
-  indexFromSlug,
   activeStage,
-  stageSlug,
-  taskSlug,
-  isDev,
   activeTaskInStage,
   activeTask,
-  issueSlug,
 } from "@/utils";
 import IssueBanner from "./IssueBanner.vue";
 import IssueHighlightPanel from "./IssueHighlightPanel.vue";
@@ -152,34 +128,18 @@ import TaskCheckBar from "./TaskCheckBar.vue";
 import type {
   Issue,
   IssueCreate,
-  IssuePatch,
-  PrincipalId,
   Database,
   Instance,
-  Stage,
-  StageId,
-  TaskId,
   Task,
   TaskDatabaseSchemaUpdatePayload,
-  StageCreate,
   TaskCreate,
-  Project,
   MigrationType,
-  TaskPatch,
 } from "@/types";
-import {
-  defaultTemplate,
-  templateForType,
-  InputField,
-  OutputField,
-} from "@/plugins";
+import { defaultTemplate, templateForType } from "@/plugins";
 import {
   featureToRef,
-  useCurrentUser,
   useDatabaseStore,
   useInstanceStore,
-  useIssueStore,
-  useIssueSubscriberStore,
   useProjectStore,
   useTaskStore,
 } from "@/store";
@@ -190,6 +150,7 @@ import {
   StandardModeProvider,
   TaskTypeWithStatement,
   IssueLogic,
+  useBaseIssueLogic,
 } from "./logic";
 
 const props = defineProps({
@@ -207,18 +168,27 @@ const emit = defineEmits<{
   (e: "status-changed", eager: boolean): void;
 }>();
 
-const router = useRouter();
 const route = useRoute();
 
-const currentUser = useCurrentUser();
-const issueStore = useIssueStore();
-const issueSubscriberStore = useIssueSubscriberStore();
 const taskStore = useTaskStore();
 const projectStore = useProjectStore();
 const databaseStore = useDatabaseStore();
 
 const create = computed(() => props.create);
 const issue = computed(() => props.issue);
+
+const {
+  project,
+  isTenantMode,
+  isGhostMode,
+  createIssue,
+  selectedStage,
+  selectedTask,
+  selectStageOrTask,
+  selectTask,
+  taskStatusOfStage,
+  isValidStage,
+} = useBaseIssueLogic({ issue, create });
 
 const issueLogic = ref<IssueLogic>();
 
@@ -231,73 +201,6 @@ watchEffect(() => {
 const issueTemplate = computed(
   () => templateForType(props.issue.type) || defaultTemplate()
 );
-
-const project = computed((): Project => {
-  if (props.create) {
-    return projectStore.getProjectById((props.issue as IssueCreate).projectId);
-  }
-  return (props.issue as Issue).project;
-});
-
-const updateAssigneeId = (newAssigneeId: PrincipalId) => {
-  if (props.create) {
-    (props.issue as IssueCreate).assigneeId = newAssigneeId;
-  } else {
-    patchIssue({
-      assigneeId: newAssigneeId,
-    });
-  }
-};
-
-const updateEarliestAllowedTime = (newEarliestAllowedTsMs: number) => {
-  if (props.create) {
-    if (isGhostMode.value) {
-      // In gh-ost mode, when creating an issue, all sub-tasks in a stage
-      // share the same earliestAllowedTs.
-      // So updates on any one of them will be applied to others.
-      // (They can be updated independently after creation)
-      const taskList = selectedStage.value.taskList as TaskCreate[];
-      taskList.forEach((task) => {
-        task.earliestAllowedTs = newEarliestAllowedTsMs;
-      });
-    } else {
-      selectedTask.value.earliestAllowedTs = newEarliestAllowedTsMs;
-    }
-  } else {
-    const taskPatch: TaskPatch = {
-      earliestAllowedTs: newEarliestAllowedTsMs,
-    };
-    patchTask((selectedTask.value as Task).id, taskPatch);
-  }
-};
-
-const addSubscriberId = (subscriberId: PrincipalId) => {
-  issueSubscriberStore.createSubscriber({
-    issueId: (props.issue as Issue).id,
-    subscriberId,
-  });
-};
-
-const removeSubscriberId = (subscriberId: PrincipalId) => {
-  issueSubscriberStore.deleteSubscriber({
-    issueId: (props.issue as Issue).id,
-    subscriberId,
-  });
-};
-
-const updateCustomField = (field: InputField | OutputField, value: any) => {
-  if (!isEqual(props.issue.payload[field.id], value)) {
-    if (props.create) {
-      props.issue.payload[field.id] = value;
-    } else {
-      const newPayload = cloneDeep(props.issue.payload);
-      newPayload[field.id] = value;
-      patchIssue({
-        payload: newPayload,
-      });
-    }
-  }
-};
 
 const runTaskChecks = (task: Task) => {
   taskStore
@@ -312,174 +215,8 @@ const runTaskChecks = (task: Task) => {
     });
 };
 
-const patchIssue = (
-  issuePatch: IssuePatch,
-  postUpdated?: (updatedIssue: Issue) => void
-) => {
-  issueStore
-    .patchIssue({
-      issueId: (props.issue as Issue).id,
-      issuePatch,
-    })
-    .then((updatedIssue) => {
-      // issue/patchIssue already fetches the new issue, so we schedule
-      // the next poll in NORMAL_POLL_INTERVAL
-      // pollIssue(NORMAL_POLL_INTERVAL);
-      emit("status-changed", false);
-      if (postUpdated) {
-        postUpdated(updatedIssue);
-      }
-    });
-};
-
-const patchTask = (
-  taskId: TaskId,
-  taskPatch: TaskPatch,
-  postUpdated?: (updatedTask: Task) => void
-) => {
-  taskStore
-    .patchTask({
-      issueId: (props.issue as Issue).id,
-      pipelineId: (props.issue as Issue).pipeline.id,
-      taskId,
-      taskPatch,
-    })
-    .then((updatedTask) => {
-      // For now, the only task/patchTask is to change statement, which will trigger async task check.
-      // Thus we use the short poll interval
-      emit("status-changed", true);
-      if (postUpdated) {
-        postUpdated(updatedTask);
-      }
-    });
-};
-
-const createIssue = (issue: IssueCreate) => {
-  // Set issue.pipeline and issue.payload to empty
-  // because we are no longer passing parameters via issue.pipeline
-  // we are using issue.createContext instead
-  delete issue.pipeline;
-  issue.payload = {};
-
-  issueStore.createIssue(issue).then((createdIssue) => {
-    // Use replace to omit the new issue url in the navigation history.
-    router.replace(`/issue/${issueSlug(createdIssue.name, createdIssue.id)}`);
-  });
-};
-
 const currentPipelineType = computed((): PipelineType => {
   return pipelineType(props.issue.pipeline!);
-});
-
-const selectedStage = computed((): Stage | StageCreate => {
-  const stageSlug = router.currentRoute.value.query.stage as string;
-  const taskSlug = router.currentRoute.value.query.task as string;
-  // For stage slug, we support both index based and id based.
-  // Index based is used when creating the new task and is the one used when clicking the UI.
-  // Id based is used when the context only has access to the stage id (e.g. Task only contains StageId)
-  if (stageSlug) {
-    const index = indexFromSlug(stageSlug);
-    if (index < props.issue.pipeline!.stageList.length) {
-      return props.issue.pipeline!.stageList[index];
-    }
-    const stageId = idFromSlug(stageSlug);
-    const stageList = (props.issue as Issue).pipeline.stageList;
-    for (const stage of stageList) {
-      if (stage.id == stageId) {
-        return stage;
-      }
-    }
-  } else if (!props.create && taskSlug) {
-    const taskId = idFromSlug(taskSlug);
-    const stageList = (props.issue as Issue).pipeline.stageList;
-    for (const stage of stageList) {
-      for (const task of stage.taskList) {
-        if (task.id == taskId) {
-          return stage;
-        }
-      }
-    }
-  }
-  if (props.create) {
-    return props.issue.pipeline!.stageList[0];
-  }
-  return activeStage((props.issue as Issue).pipeline);
-});
-
-const selectStageOrTask = (
-  stageId: StageId,
-  taskSlug: string | undefined = undefined
-) => {
-  const stageList = props.issue.pipeline!.stageList;
-  const index = stageList.findIndex((item, index) => {
-    if (props.create) {
-      return index === stageId;
-    }
-    return (item as Stage).id == stageId;
-  });
-  router.replace({
-    name: "workspace.issue.detail",
-    query: {
-      ...router.currentRoute.value.query,
-      stage: stageSlug(stageList[index].name, index),
-      task: taskSlug,
-    },
-  });
-};
-
-const selectTaskId = (taskId: TaskId) => {
-  const taskList = selectedStage.value.taskList as Task[];
-  const task = taskList.find((t) => t.id === taskId);
-  if (!task) return;
-  const slug = taskSlug(task.name, task.id);
-  const stage = selectedStage.value as Stage;
-  selectStageOrTask(stage.id, slug);
-};
-
-const selectTask = (task: Task) => {
-  if (!create.value) return;
-
-  const stage = (issue.value as Issue).pipeline?.stageList.find(
-    (t) => t.id === task.id
-  );
-  if (!stage) {
-    return;
-  }
-  const slug = taskSlug(task.name, task.id);
-  selectStageOrTask(stage.id, slug);
-};
-
-const selectedTask = computed((): Task | TaskCreate => {
-  const taskSlug = route.query.task as string;
-  const { taskList } = selectedStage.value;
-  if (taskSlug) {
-    const index = indexFromSlug(taskSlug);
-    if (index < taskList.length) {
-      return taskList[index];
-    }
-    const id = idFromSlug(taskSlug);
-    for (let i = 0; i < taskList.length; i++) {
-      const task = taskList[i] as Task;
-      if (task.id === id) {
-        return task;
-      }
-    }
-  }
-  return taskList[0];
-});
-
-const isTenantMode = computed((): boolean => {
-  if (project.value.tenantMode !== "TENANT") return false;
-  return (
-    props.issue.type === "bb.issue.database.schema.update" ||
-    props.issue.type === "bb.issue.database.data.update"
-  );
-});
-
-const isGhostMode = computed((): boolean => {
-  if (!isDev()) return false;
-
-  return props.issue.type === "bb.issue.database.schema.update.ghost";
 });
 
 const selectedMigrateType = computed((): MigrationType => {
@@ -492,20 +229,6 @@ const selectedMigrateType = computed((): MigrationType => {
     ).migrationType;
   }
   return "MIGRATE";
-});
-
-const allowEditSidebar = computed(() => {
-  // For now, we only allow assignee to update the field when the issue
-  // is 'OPEN'. This reduces flexibility as creator must ask assignee to
-  // change any fields if there is typo. On the other hand, this avoids
-  // the trouble that the creator changes field value when the creator
-  // is performing the issue based on the old value.
-  // For now, we choose to be on the safe side at the cost of flexibility.
-  return (
-    props.create ||
-    ((props.issue as Issue).status == "OPEN" &&
-      (props.issue as Issue).assignee?.id == currentUser.value.id)
-  );
 });
 
 const showPipelineFlowBar = computed(() => {
@@ -572,29 +295,6 @@ const supportBackwardCompatibilityFeature = computed((): boolean => {
   return engine === "MYSQL" || engine === "TIDB";
 });
 
-const taskStatusOfStage = (stage: Stage | StageCreate) => {
-  if (props.create) {
-    return stage.taskList[0].status;
-  }
-  const activeTask = activeTaskInStage(stage as Stage);
-  return activeTask.status;
-};
-
-const isValidStage = (stage: Stage | StageCreate) => {
-  if (!props.create) {
-    return true;
-  }
-
-  for (const task of stage.taskList) {
-    if (TaskTypeWithStatement.includes(task.type)) {
-      if (isEmpty((task as TaskCreate).statement)) {
-        return false;
-      }
-    }
-  }
-  return true;
-};
-
 // Determine which type of IssueLogicProvider should be used
 const logicProviderType = computed(() => {
   if (isTenantMode.value) return TenantModeProvider;
@@ -613,13 +313,14 @@ watch(
     }
   }
 );
+
 const onStatusChanged = (eager: boolean) => emit("status-changed", eager);
 
 provideIssueLogic(
   {
     create,
     issue,
-    project: project,
+    project,
     template: issueTemplate,
     selectedStage,
     selectedTask,
@@ -633,8 +334,6 @@ provideIssueLogic(
     selectStageOrTask,
     selectTask,
     onStatusChanged,
-    patchTask,
-    patchIssue,
     createIssue,
   },
   true
