+++ frontend/src/components/Issue/logic/index.ts
@@ -4,6 +4,7 @@ import TenantModeProvider from "./TenantModeProvider";
 import GhostModeProvider from "./GhostModeProvider";
 import StandardModeProvider from "./StandardModeProvider";
 
+export * from "./base";
 export * from "./common";
 export * from "./extra";
